minimal_calendar/requirements.txt
absl-py==1.3.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.8.1
asttokens==2.2.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==22.2.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.9.0
cachetools==5.2.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.2.0
charset-normalizer==2.1.1
click==8.1.7
cloup==3.0.5
comm==0.1.2
contourpy==1.0.6
cryptography==44.0.0
cycler==0.11.0
dataclasses-json==0.6.7
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
Django==5.1.5
django-cors-headers==4.6.0
django-crispy-forms==1.14.0
django-froala-editor==4.1.3
django-recaptcha==3.0.0
django-tinymce==3.6.1
Django-Verify-Email==2.0.3
djangorestframework==3.14.0
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
entrypoints==0.4
et-xmlfile==1.1.0
eval_type_backport==0.2.2
executing==1.2.0
faiss-cpu==1.10.0
fastapi==0.115.8
fastjsonschema==2.16.2
ffmpeg==1.4
filelock==3.17.0
Flask==3.1.1
flatbuffers==22.12.6
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.2.0
gast==0.4.0
glcontext==2.5.0
google-auth==2.15.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.51.1
gTTS==2.5.4
gunicorn==23.0.0
h11==0.14.0
h5py==3.7.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.28.1
idna==3.4
ipykernel==6.19.4
ipython==8.7.0
ipython-genutils==0.2.0
isoduration==20.11.0
isosurfaces==0.1.2
itsdangerous==2.2.0
jax==0.6.1
jaxlib==0.6.1
jedi==0.18.2
Jinja2==3.1.2
jiter==0.8.2
joblib==1.2.0
json5==0.9.10
jsonpatch==1.33
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.1
jupyter_server==2.0.5
jupyter_server_terminals==0.4.3
jupyterlab==3.5.2
jupyterlab-pygments==0.2.2
jupyterlab_server==2.17.0
keras==2.11.0
kiwisolver==1.4.4
langchain==0.3.18
langchain-community==0.3.17
langchain-core==0.3.35
langchain-text-splitters==0.3.6
langsmith==0.3.8
libclang==14.0.6
manim==0.18.1
ManimPango==0.5.0
mapbox-earcut==1.0.1
Markdown==3.4.1
markdown-it-py==3.0.0
MarkupSafe==2.1.1
marshmallow==3.26.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapipe==0.10.21
mistune==2.0.4
ml_dtypes==0.5.1
moderngl==5.10.0
moderngl-window==2.4.6
mpmath==1.3.0
multidict==6.1.0
multipledispatch==1.0.0
mypy-extensions==1.0.0
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
networkx==3.3
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
openai==1.62.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
openpyxl==3.0.10
opt-einsum==3.3.0
orjson==3.10.15
packaging==24.2
pandas==1.5.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==11.1.0
platformdirs==2.6.0
playsound==1.3.0
plyer==2.1.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
propcache==0.2.1
protobuf==3.17.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==19.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.26.0
pycparser==2.21
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydub==0.25.1
pygame==2.6.1
pyglet==2.0.15
Pygments==2.13.0
PyJWT==2.10.1
pyparsing==3.0.9
PyQt5==5.15.11
PyQt5-Qt5==5.15.16
PyQt5_sip==12.17.0
PyQt6==6.9.0
PyQt6-Qt6==6.9.0
PyQt6_sip==13.10.0
pyrr==0.10.3
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.4
python3-openid==3.2.0
pyttsx3==2.98
pytz==2022.7
PyYAML==6.0
pyzmq==24.0.1
regex==2024.11.6
reportlab==4.4.1
requests==2.32.3
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rsa==4.9
safetensors==0.5.2
scikit-learn==1.2.0
scipy==1.13.0
screeninfo==0.8.1
Send2Trash==1.8.0
sentence-transformers==3.4.1
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
skia-pathops==0.8.0.post1
sniffio==1.3.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sounddevice==0.5.2
soupsieve==2.3.2.post1
SQLAlchemy==2.0.38
sqlparse==0.4.2
srt==3.5.3
stack-data==0.6.2
starlette==0.45.3
svgelements==1.9.6
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.29.0
termcolor==2.1.1
terminado==0.17.1
threadpoolctl==3.1.0
tiktoken==0.8.0
tinycss2==1.2.1
together==1.4.0
tokenizers==0.21.0
tomli==2.0.1
torch==2.6.0
tornado==6.2
tqdm==4.66.4
traitlets==5.8.0
transformers==4.48.3
triton==3.2.0
typer==0.15.1
typing-inspect==0.9.0
typing_extensions==4.12.2
uri-template==1.2.0
urllib3==1.26.13
uvicorn==0.34.0
watchdog==4.0.0
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==3.1.3
whitenoise==6.9.0
wrapt==1.14.1
xlrd==2.0.1
xlwt==1.3.0
yarl==1.18.3
zstandard==0.23.0
